+++ crates/rerun/src/lib.rs
@@ -1,5 +1,8 @@
 //! The main Rerun library and binary.
 //!
+//! ## Examples
+//! See <https://github.com/rerun-io/rerun/tree/latest/examples/rust>.
+//!
 //! ## Library
 //! See [`Session`] and [`MsgSender`].
 //!