+++ rerun_py/rerun_sdk/rerun/log/arrow.py
@@ -57,6 +57,10 @@ def log_arrow(
         The entity is not time-dependent, and will be visible at any time point.
 
     """
+
+    if not bindings.is_enabled():
+        return
+
     instanced: Dict[str, Any] = {}
     splats: Dict[str, Any] = {}
 
@@ -81,9 +85,10 @@ def log_arrow(
     if ext:
         _add_extension_components(instanced, splats, ext, None)
 
-    if instanced:
-        bindings.log_arrow_msg(entity_path, components=instanced, timeless=timeless)
-
     if splats:
         splats["rerun.instance_key"] = InstanceArray.splat()
         bindings.log_arrow_msg(entity_path, components=splats, timeless=timeless)
+
+    # Always the primary component last so range-based queries will include the other data. See(#1215)
+    if instanced:
+        bindings.log_arrow_msg(entity_path, components=instanced, timeless=timeless)