+++ rerun_py/rerun_sdk/rerun/log/extension_components.py
@@ -110,6 +110,10 @@ def log_extension_components(
         If true, the components will be timeless (default: False).
 
     """
+
+    if not bindings.is_enabled():
+        return
+
     identifiers_np = np.array((), dtype="uint64")
     if identifiers:
         try:
@@ -126,9 +130,10 @@ def log_extension_components(
 
     _add_extension_components(instanced, splats, ext, identifiers_np)
 
-    if instanced:
-        bindings.log_arrow_msg(entity_path, components=instanced, timeless=timeless)
-
     if splats:
         splats["rerun.instance_key"] = InstanceArray.splat()
         bindings.log_arrow_msg(entity_path, components=splats, timeless=timeless)
+
+    # Always the primary component last so range-based queries will include the other data. See(#1215)
+    if instanced:
+        bindings.log_arrow_msg(entity_path, components=instanced, timeless=timeless)