+++ rerun_py/rerun_sdk/rerun/log/points.py
@@ -42,7 +42,7 @@ def log_point(
     timeless: bool = False,
 ) -> None:
     """
-    Log a 2D or 3D point, with optional color.
+    Log a 2D or 3D point, with a positions and optional colors, radii, label, etc.
 
     Logging again to the same `entity_path` will replace the previous point.
 
@@ -84,6 +84,10 @@ def log_point(
         If true, the point will be timeless (default: False).
 
     """
+
+    if not bindings.is_enabled():
+        return
+
     if keypoint_id is not None and class_id is None:
         class_id = 0
     if position is not None:
@@ -118,12 +122,13 @@ def log_point(
     if ext:
         _add_extension_components(instanced, splats, ext, None)
 
-    if instanced:
-        bindings.log_arrow_msg(entity_path, components=instanced, timeless=timeless)
-
     if splats:
         splats["rerun.instance_key"] = InstanceArray.splat()
         bindings.log_arrow_msg(entity_path, components=splats, timeless=timeless)
+
+    # Always the primary component last so range-based queries will include the other data. See(#1215)
+    if instanced:
+        bindings.log_arrow_msg(entity_path, components=instanced, timeless=timeless)
 
 
 def log_points(
@@ -140,7 +145,7 @@ def log_points(
     timeless: bool = False,
 ) -> None:
     """
-    Log 2D or 3D points, with optional colors.
+    Log 2D or 3D points, with positions and optional colors, radii, labels, etc.
 
     Logging again to the same `entity_path` will replace all the previous points.
 
@@ -185,6 +190,10 @@ def log_points(
         If true, the points will be timeless (default: False).
 
     """
+
+    if not bindings.is_enabled():
+        return
+
     if keypoint_ids is not None and class_ids is None:
         class_ids = 0
     if positions is None:
@@ -244,8 +253,9 @@ def log_points(
     if ext:
         _add_extension_components(comps[0], comps[1], ext, identifiers_np)
 
-    bindings.log_arrow_msg(entity_path, components=comps[0], timeless=timeless)
-
     if comps[1]:
         comps[1]["rerun.instance_key"] = InstanceArray.splat()
         bindings.log_arrow_msg(entity_path, components=comps[1], timeless=timeless)
+
+    # Always the primary component last so range-based queries will include the other data. See(#1215)
+    bindings.log_arrow_msg(entity_path, components=comps[0], timeless=timeless)